Fix vanilla-map NPC loading and give-item crash on 64-bit
Three 64-bit pointer/width mismatches in legacy map-load and dialogue
paths broke the Omerta intro on macOS/Linux. None change the on-disk
map format - all are read/in-memory layout fixes only.

1. OLD_SOLDIERCREATE_STRUCT_101 (vanilla v5.0 detailed placements):
   CHAR16 name[10] is 4 bytes/char in memory (2 on disk) and
   SOLDIERTYPE* pExistingSoldier is 8 bytes (4 on disk), so the raw
   blob read scrambled every field after `name`, derailing
   LoadSoldiersFromMap at the first civilian placement -> no NPCs in
   A9 (kid, Fatima, etc.). Make the struct's POD region byte-for-byte
   the 32-bit on-disk layout (UINT16 name, UINT32 pointer slot) and
   widen/narrow at the conversion operators. Matches the existing
   OLD_MERCPROFILESTRUCT_101 / MAPDISK_SOLDIERCREATE_STRUCT precedent.

2. _OLD_SCHEDULENODE / _OLD_SCHEDULENODE_PRE_ITS (v<8.0 NPC schedules):
   the leading `next` pointer was 8 bytes in memory vs 4 on disk,
   shifting every schedule field by 4. A9 has MAP_NPCSCHEDULES_SAVED
   set, so this broke right after the soldier fix. Make `next` a
   UINT32 slot (it's a transient link rebuilt on load).

3. Give-item crash: the conversation system smuggles an OBJECTTYPE*
   through uiApproachData, but that carrier was UINT32. On 64-bit the
   pointer was truncated at the InitiateConversation call site and
   crashed when cast back in Converse/ReturnItemToPlayerIfNecessary.
   Widen uiApproachData to uintptr_t end-to-end (same approach the
   codebase already uses for HWFILE) and drop the truncating cast.

Verified: full Omerta intro now works - kid + Fatima spawn in A9,
giving Fatima the letter no longer crashes, and the escort through
A10 and the A10 basement (v7.0 map) loads and recruits correctly.

+++ Tactical/Handle Items.cpp
@@ -5057,7 +5057,7 @@ void SoldierGiveItemFromAnimation( SOLDIERTYPE *pSoldier )
 //			}
 
 			// Now intiate conv
-			InitiateConversation( pTSoldier, pSoldier, APPROACH_GIVINGITEM, (UINT32)(uintptr_t)&gTempObject );
+			InitiateConversation( pTSoldier, pSoldier, APPROACH_GIVINGITEM, (uintptr_t)&gTempObject );
 		}
 	}
